Backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
safetensors
python-multipart
